pyproject.toml
[project]
name = "web-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "bedantH", email = "bedanthota@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "pika>=1.3.2",
    "sqlmodel>=0.0.31",
]

[project.scripts]
web-crawler = "web_crawler:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
frontier = { workspace = true }
slave = { workspace = true }
master = { workspace = true }

[tool.uv.workspace]
members = [
    "services/*",
    "shared",
]

[tool.uv]
dev-dependencies = [
    "frontier",
    "slave",
    "master",
]
package = false
